This represents a HUGE number of changes. They mainly deal with converting from the first DANA incarnation (CVS tag DANA-0-1) to a new template-based one. There may be other unrelated changes, but given the number of files being committed here, it would take too long to commit them all individually.
As part of the modification, the DContainer class has been dropped. The
template system requires that all data be kept in classes (with DContainer,
it was all kept in C-style structs underneath). It was also natural to move
to a naming convention in which the factory name is based on the name of
the data class rather than the container's name. In other words, files
like DFactory_FCALHits.cc were renamed to DFactory_DFCALHit.cc.

The functionality of DContainer has been replaced by the use of STL vectors.

This has been checked to compile and link on Mac OS X, Linux RHEL3, and
SunOS using the GNU 3 compiler. It has also been compiled and linked
on SunOS using the Sun Compiler, though there are segmentation faults
when running that version.