This is a merge of the code developed in the jana-branch branch. It represepents the divorce of the orginal DANA code into the joint JANA project. The JANA code officially resides in another repository (/group/12gev_phys/svnroot).
The command used to merge the changes from the jana-branch was:

svn merge -r 1855:1898 svn+ssh://svn.jlab.org/group/halld/Repositories/svnroot/branches/jana-branch

Note that in the future, the jana-branch cranch may be deleted from the repository for the sake of tidiness. It should always be available through the subversion archive though.

Note also that there is still a DANA directory. This has things like the Hall-D sub-class of JApplication where we put Hall-D specific behaviour.