This fixes a long standing problem that caused the build using the SunOS compiler to fail. The GNU C++ compiler worked fine on this code previously which is why it was allowed to linger so long.
The problem seems to be that the SunOS compiler doesn't like "const" types to be used for the key type in std::map types. This was a little tough to get from the error message which spoke of multiple definitions. Anyway, changing the key type from "const int" to just "int" everywhere for the one map used seems to have fixed the problem. It now compiles cleanly on SunOS and still does on Linux.