Fixed what appeared to be a bug in U_OF_WIRE_ZERO and U_OF_STRIP_ZERO where the 1 was being subtracted after multiplying by STRIP_SPACING.
Changed the strip spacing to 0.5cm and the number of strips from 192 to 216
after speaking with Simon. These are very soft numbers and could change at
anytime, but at least they are the most current.

Fixed a bug where the index of one of xoutlocal was off by one. This caused a problem with the calculation of avalanche_y.