This modifies the use of the jparms global object for accessing configuration parameters. It corresponds to a change in the JANA package. The comment from that change is here:
This changes the way the global JParameterManager object is implemented. Previously, it was simply a global scope object called jparms allocated on the stack prior to main getting called. This caused problems with plugins since the plugins (on some OSes) would link in a separate jparms (very bad). 

Now, jparms has been replaced with gPARMS which is a pointer to a JParameterManager object allocated on the heap in the JApplication constructor. Also added was a global routine called InitJANAPlugin(...) that should be called by all plugins right away with the JApplication pointer as the only argument. This sets the gPARMS point to the one kept in the JApplication object. It also provides a place for any other bootstrapping needed to ensure globals on both sides of the DLL are aligned.