These 2 files (grkuta.cc and gufld.F) represent the magnetic filed stepper and map routines used by hdgeant respectively. The grkuta.cc file was converted from the GEANT3.2.1 file grkuta.F using f2c and then modified to use the DMagneticFieldMap class to access the magnetic field.
The gufld.F file was cut from the generated file hddsGeant3.F. It was modified to always return 3 from getmap. This will be converted to C++ at a future time. right now, it is included so that the interpolation will be done in the same way as  hdgeant does it. Note that this now requires the presence of solenoid.map.
-- This line, and those below, will be ignored--

A    grkuta.cc
A    gufld.F