Modified HDGeant to still produce hits in the CDC and FDC even if energy loss is turned off. Energy loss is controlled via the "LOSS" card in control.in.
This allows us to get digitized values from pristine tracks. It is very un-physical, but allows checking of the magnetic field swimmer in DANA at a level otherwise not accessible. 

To get the old LOSS=0 behavior (why would you ever want it??) set LOSS to -1. 

This should not affect normal simulations where LOSS should be set to 1.