This represents a few semi-major changes to the tracking package. There is still a problem with multi-track events, but I want to backup the current state of the code since it represents a fair amount of work. At this point it looks like we are still fairly susceptable to local minima which still needs to be addressed.
One of the major changes is that the fitter now fits x and v directions in the rest frame of the initial reference trajectory point rather than x and y. The v direction is defined to be perpendicualr to the momentum (as is x) where as y was not necessarily perpendicular.

Also, backtracking was implemented to try and salvage fits where we "overstep" the correct point in parameter space.

A routine for projecting a reference trajectory up to a plane was written, but is still buggy.