Added new *temporary* factory for BCAL reconstruction.
This is a clone of the Regina algorithm (which seems to be the best we have to date for BCAL reconstruction), but it runs about an order of magnitude faster.  It is being added to the repository so it can be checked/compared by others since the changes from the original Regina code are rather dramatic in some places.  The output of this factory should be identical to that that comes out of the default factory.  

Eventually this factory will get tuned so it is no longer a clone of the old algorithm and then it will replace the default DBCALShower factory.  

To use this factory provide the tag "IU" when getting the DBCALShower vector.