Modified data model for "product" to include:
pdgtype
id
parentid
mech

This is sufficient to describe the heritage of all particles in
the intermediary and final states. Even the ones that are not
to be tracked by geant. This is motivated by the output of pythia.

Changes were made to both the XML and schema files and updated 
versions of the generated files hddm_s.c and hddm_s.h are being
committed as well.

The DEventSourceHDDM.cc file was modified pass the changes on
to DMCThrown (which I just realized is not in the current commit
list and so will be committed in the next revision).

The genr8_2_hddm program and hddmInput.c file of HDGeant were modified
to fill in zeros for the new values to avoid their being filled
with random values. Finally, pythiagen was modified to fill in the
new fields with the output from pythia.