Updated thrown and reconstructed table values. The reconstructed values work from DKinematicData objects so one can display either DTrackCandidate or DTrack objects there.
Made the default setting for both the graphic and table display of track parameters be for DTrackCandidate:LINK since that is the factory currently under development.

Disabled drawing of FCAL. The drawing of 2800 polygons makes it much slower and that is unnecessary at the moment. It will be turned back on later with possibly a new paradigm for drawing to make it more efficient.