Added p_trans and phi members to DRiemannFit class. Added GuessChargeFromCircleFit and FitCircleRiemann methods to DQuickfit.
The FitCircleRiemann method creates a temporary DRiemannFit
object to do the fit and all of the hits are copied into it
so it is not the most efficient way of
doing things. It does make it easy to switch between the
DQuickFit method (circle must pass through origin) and the
Riemann fit method (origin is included as a separate point
with no strict requirements.) though.