Updated number of straws and radii to be consistent with the current default for CDC in hdds (option-1).
Also, the sign of the stereo angle was flipped. This is needed to
make things work empirically. I recall there was an issue before
with hdds actually having it backward compared to the documentation
and since the recontruction followed the documentation, it all worked.
Now, hdds is consistent with the documentation for the V4 geometry
so I have to make the explicit flip. Go figure.