These changes allow the magnetic field map to be obtained through the calibration database API rather than from the local solenoid.map file. The changes were originally made 6 months ago in a private area of the repository but never copied over. The main trunk has since started to diverge from the development version. Also, this will allow us to make the 1500A field the standard which is long overdue.
The most significant change here is that the executables must now be
linked with the HDGEOMETRY library and JANA which contain C++ code.
As such, "main()" must now be defined in a C++ file and the "PROGRAM"s
redefined as SUBROUTINEs and called from their respective C++
counterparts.

Note that this change does not yet actually use the field from the
calibDB. This is because hddsGeant3.F needs to be modified to call
it. That change will be made on a subsequent commit.