Added code from Elke to allow Ko decays (among others) to be turned on in pythia so we can smear the products parametrically for resolution studies.
By default, the decays are turned off to allow GEANT to decay the Ko.
To turn it on, uncomment the call to pythia6_setdecay().

Note that this program is considered obsolete by its author "E. Chudakov" since its functionality is fully contained in bggen now. This particular change should be copied over to that program at some point, but for now, this is where it is so this is where it gets committed to.