Modified makefiles to allow "make pristine" to work from the topmost makefile.
This includes a changed to BMS/Makefile.bin to execute a "set nonomatch"
just before the "rm -rf ..." statement in the clean target. This was to
address a problem that suddenly appeared in the HDGeant directory where
it was failing due to a "no match" error. The set nonomatch is for tcsh,
but seems innocuous to bash so hopefully won't cause any problems there.

For several of the Makefiles modified, the pristine target was simply
added with a dependancy on clean.