Updated resolutions tables to adjust momentum range from 0 to 7 GeV/c (was 0.2 to 7.1) and angle range from 0 to 160 degrees. These match up with what was thrown in the simulation.
At the same time, adjusted number of bins on all histograms (resolutions
and efficiency) to be the same. A minor error caused the dp_over_p
histos to have a different bin definition than the other histograms binned 
by ptot and theta.

Added a check in DTrackingResolutionGEANT::GetResolution to make sure
requested values are in range of histograms. If not, resolutions of
zero are returned.