- Only call FDPREE if HADR flag is set to 3 or more indicating FLUKA is being used.
This is motivated by the situation on a Fedora8 machine with the default cernlib 2006
installed which does not have FLUKA support. When a value of HADR=1 is specified
in the control.in file, hdgeant would quit on this call complaining about
the absence of FLUKA. A value of HADR=1 should indicate only GEISHA be used for
hadronic interactions so the FLUKA specific call FDPREE should not be made.

An interesting side note: When a value of HADR=3 was specified in control.in, it 
seemed to set IHADR to 1. For all other values, IHADR was set to whatever the HADR
flag was set to. This may have been done in the same dummy library that replaced the
FLUKA calls since HADR=3 means GEISHA and FLUKA and HADR=1 means only GEISHA. The end
result is that setting HADR=3 is equivalent to HADR=1 on systems without FLUKA support
compiled in, but with no warning messages to indicate to the user this is happening.
Using a value of HADR=4 will cause the program to quit with an appropriate error
message.