Implemented work-around for problem introduced in ROOT 5.16 or 5.17 where the TGComboBox objects would not properly display the drop-down menus for the factory tags of DTrackCandidates, DTracks, etc..
The work-around was mainly to initialize the menu's with 100 entries when initially created. The contents are overwritten anyway so the initial values are never seen by the user. 

In addition, some bugs were fixed in the filling of DParticle tag's menu and restoring that value from the preferences file.

The DTrackCandidate color was hardwired to 30.