Fixed calculation of residual based tracking chi-squared. Now the chi-squared value and number of degrees of freedom are correctly calculated.
Changed the error used in wire-based fits to be the half-cell size over sqrt(12). For the CDC, this was 0.8/sqrt(12) and for the FDC it was 0.5/sqrt(12). These are consistent with the emprically determined values used before.

Set the CDC_SIGMA value in the ALT2 fitter to 150 microns to be consistent with the ALT1 setting. (Both should really not be setting the default!)