Added plugins directory to Simulations.
Added hdparsim plugin to Simulations/plugins.

This plugin re-implements the HDParSim program as a plugin
that produces DParticle objects and DPhoton objects both
with the HDParSim tag. This will allow the same tree/hist
maker programs to be used with parametrically simulated data
using a DANA program and the output of a physics generator
(HDDM format).