Modify how FCAL smearing data is transported from hdgeant through the reconstruction.
hdgeant now fills the FcalTruthHit structures (instead of the FcalHit structures). The mcsmear program copies the values from FcalTruthHits to FcalHits simultaneously smearing the energy and time of indivdual hits and applying a threshold cut on the energy.

This corrects a poor design choice made earlier in which the truth information was carried from hdgeant in FcalHits and then mcsmear copied it into FcalTruthHits by mcsmear. Now, the truth info always exists only in the branch with "truth" in the name.