Install our own Error Handler for ROOT error messages. This will allow us to supress those annoying "Error in <TDecompLU::DecomposeLUCrout>:" type messages.
The ROOT system has a level associated with the errors. It appears that the above message (most others) have a level of 3000. By default, the supression level is set to only print messages with a level >10000 or if the abort flag is set on the message.

One can set the level for error messages are supressed by setting the ROOT_ERROR_LEVEL_SUPRESS config. parameter. To get the old behavior in which ROOT's default handler is always called, pass

-PROOT_ERROR_LEVEL_SUPRESS=-1 on the command line.