Defer instantiation of magnetic field map, Lorentz deflection table, and geometry objects until requested. This avoids the costly reading-in of these tables when they are not needed.
A consequence of this is that the time it takes to read in the map is 
counted during the event processing time which may (and often did) 
cause the JANA heartbeat monitor to kill the thread after 8 seconds
believing it to be hung. Thus, we also set the timeout to 30 seconds
to accomodate that. This will cause the THREAD_TIMEOUT config. parameter
to now always be displayed for Hall-D programs.