Expanded position cuts for grouping hits between planes to take into account resolution effects. This results in a significantly increased number of reconstructed DTOFPoints for tracks hitting near the edges of the paddles.
Improved the hit-matching algorithm: could have had a single DTOFHit matched into several DTOFPoints, especially with the increased cut width (e.g. hits in adjacent paddles).  
DTOFPoints are now reconstructed by looping: find the closest matching pair of hits, build the object, remove any remaining matches using these hits.  

Also, an algorithm was introduced to reconstruct the properties of the DTOFPoint objects if one of the PMT signals in one or both of the planes was below the energy threshold. 
However, this algorithm will not be activated until the rest of the system is ready (e.g. grouping/ignoring simultaneous low-energy reconstructed hits in adjacent paddles).