Added place in HDDM to store random number seeds used in simulation. Places were added to hold the 2 seeds from GEANT and the one seed (anticipated) for mcsmear.
HDGeant was modified to look for random seeds in the input file (if one is used) and use those, if found, to seed the random number generator at the start of each event. The seeds used (whatever their source) are recorded in the output file with the event.

The comments in the example control.in file regarding the use of RNDM were modified to inform users that the value(s) they enter there may be overridden.

This system allows one to extract a single event from an hddm file (a'la hddm_cull_events) and re-simulate in in hdgeant with the exact same seeds.

Similar functionality is not yet implemented in mcsmear.