Making the Time Spectrum method for BCAL simulation the default. The code has existed for some time, but the method set by #defines. To switch back to the old cell-based method, adjust the #defines at the top of programs/Simulation/mcsmear/smear_bcal.cc and programs/Simulation/HDGeant/hitBCal.cc accordingly.
Also included is a change to libraries/BCAL/DBCALPoint.cc that changes the scale of the amplitude values to roughly convert them from fADC units to GeV. This is a temporary change so as not to break all of the reconstruction code. If one wishes to go back to the cell method, this change should be reveresed as well.