Added ability to create a "thrown" tree, where only the thrown particle data is output (+ run & event number). This is useful for calculating the acceptance (generated data = denominator).
Also, the interface for creating and writing trees was changed. 
Tree writing was changed to reflect the new "thrown" tree option.

Tree creation was changed to separate tree filling from tree creation. 
In one program execution you may not want to save any entries to the tree at all. 
However, you cannot merge this histogram with others with the tree because the ROOT file contents will be different.
This ensures that every ROOT file will have the trees.