- Adjusted module names to be consistent with those from Fernandos spreadsheet - Replaced some of the "end" fields which used 0 and 1 to have "N", "S", ...   to be consistent with translation table generated from the spreadsheet - increased max number of crates, channels to allow for those defined in real   translation table
A change was also made to mc2codaWrite to prevent it from seg. faulting if hit
data for an undefined channel is passed. That change is in the mc2coda library
(and so, not here).

Program runs, but with lots of "unknown map entry" errors (from this coda) as
well as lots of "Too many hits" warnings (from mc2coda).