Created DDetectorMatches class to store matching information between different detector systems. Note that this is not in REST yet.
Changed the members of DChargedTrackHypothesis to include the matching information and to remove the reference trajectory. 

Added timing uncertainty variable to DTOFPoint (value not set). 

Fixed bug in setup for time-based tracking, where the input time to the match-to functions was not reset in between calls. 

Only swim new time-based tracks in the "combo" factory when needed (e.g. significantly different particle mass or charge). 

Changed many functions to interface with the new classes and methods.