Candidates: Cap maximum # track circles at 10000, reduce memory footprint by switching double -> float. To combat memory spikes. Time & Wire-Based: Increase max size of DReferenceTrajectory vector to 50, and allocate this many objects during brun().
This is because each DReferenceTrajectory swim step array needs large, free sections of contiguous memory or else it will crash
(within the recent resize, now array needs 860kB) (can be 100+ objects in an event at once). 

This was causing difficult-to-reproduce crashes on events with many tracks (typically 30+ wire-based) in limited-memory environments.