Copied core part of DAQ_WRD2PI plugin into DAQ plugin so that it can generate Df250PulseIntegral objects from Df250WindowRawData objects. Added a "pedestal" member to Df250PulseIntegral.
Configuration parameter EVIO:EMULATE_PULSE_INTEGRAL_MODE can
be used to turn off the emulation which is on by default.
Note that with emulation on, it activates it only if no
Df250PulseIntegral objects exist for the event. This could
give inconsistencies if reading both window raw data and
pulse data, but some events don't have pulse data objects.