Added emulation of Df125PulseTime objects from the Df125WindowRawData objects. For the moment, this uses a much simpler algorithm than what will be put into the FPGA. It sets up the framework though and allows the drift times to be seen in hdview2.
The conversion from drift time to distance is done in the DCDCTrackHit_factory.
Previously, this always set the "dist" value to be -1. It was to be
filled in later by the track fitter since only then would a time of flight
be known. The calculation copies the code from the DTrackFitterKalmanSIMD
class and assumes tof to wire is zero and B-field is 2T. These values
are currently hardwired and the second shoul be changed to at least use
an average field based on ring.

A shift of 170ns in the time is also hardwired for the moment in the
emulation routine for the Df125PulseTime. This is empirical from the
first 147 events of the original BCAL/CDC cosmic run data.