If a TOF hit in one plane was not matched to one in another plane, BUT it's position is well-defined: create a DTOFPoint object for it.
The TOF hit position is NOT well defined if it's on a single-ended paddle, OR if one of the PMT readout energies is below threshold. 

These could theoretically still be saved, and their position could be determined via matching to tracks in the other direction, but this has been disabled. 

It could be enabled in the future, but it requires updates to DDetectorMatches and the REST format:
 - The hit energy & time for these have to corrected due to attenuation loss and light propagation time.