Created new RF library, added to build system. Added PS & PSC to DetectorSystem_t. Added MODULE_TYPE to DF1TDCHit objects. Added extraction of RF signals to the translation table code.
Created TTabUtilities class. This class should be used to convert F1TDC "digi" times to ns. If both DCODAROCInfo and DF1TDCConfig objects are available (late 2014 and beyond), the global system clock from the each ROC's TI is used as the reference time, and the F1TDC config info is read in from the datastream (rather than the CCDB) (e.g. "tframe," etc.).  If these objects are not available (early 2014 data), then the old method (Beni-cable + CCDB constants) is used. 

Modified all hit factories using DF1TDCHit's to use TTabUtilities instead. Note that TTabUtilities must be grabbed for every event (not just in brun) because it gets and saves the DCODAROCInfo objects for each event.