Changes to hitStart.c
These changes modify the way in which the energy deposition (dE) and hit times (t) are calculated for charged tracks intersecting the Start Counter in HDGeant.  Both attenuation and propagation time correction factors, determined from bench and beam data respectively, are obtained from the CCDB and applied to the hit truth information on a paddle by paddle basis. Prior to this change, one overall attenuation length and effective velocity correction was applied to all 30 paddles over the entire scintillator geometry.