Filter out extra debug flags
Skip debug flags ("-g") in AddCompileFlags(). On CentOS6, an extra
debug flag was identified as coming from a cflag from "jana-config",
and it from a cflag from "mysql_config". This was preventing the
command-line option DEBUG=0 from taking effect on CentOS6. CentOS7
was not affected because on it, mysql is not configured with "-g".
The debug flag which is added when DEBUG=1 (in the SConstruct) is not
affected by this commit. This commit prevents extra debug flags.