Remove undefined symbol from BCAL plugins for RootSpy
There was an undefined-symbol runtime error when attaching BCAL plugins
to RootSpy for testing. The offending symbol was "typeinfo for DApplication".
A workaround was previously introduced which appended additional libraries
until all symbols were defined. While this fixed the problem at the time, it
resulted in a size increase of a few times for the shared objects, and is no
longer functioning at this time.

The undefined "typeinfo for DApplication" can be traced to the DBCALGeometry
constructor, where a pointer to DApplication is used to access the geometry
information and calculate the BCAL_PHI_SHIFT data member. As this is the only
place where DApplication is used in DBCALGeometry and BCAL_PHI_SHIFT is no
longer in use, the relevant code has been commented out.