Comment out code added for RootSpy-macro testing
Additional libraries are needed when attaching a monitoring plugin
directly to RootSpy for testing. However, this results in a very
bloated shared object file, and is unnecessary for production use
of RootSpy and usage of the plugin with hd_root/hd_ana. The required
additional libraries can be added manually in the SConscript for
testing purposes, and is necessary when testing all monitoring plugins
in this manner, not only the BCAL ones.