In dE/dx cut: Add ability to do max rejection. Now you can tell it whether to focus on keeping signal or on rejecting background.  If you choose to reject background, it will use the opposite selection cut to cut away the background.
This would effectively cut away all tracks above p = 1 GeV/c, however, it is NOT
performed for tracks with timing information. In this case, it is
assumed that the timing information is sufficient to reject the
background, and the track is treated as if the flag was false (focus on
selection).