Remove unsupported compiler flag for clang builds
This commit fixes 2 build errors introduced by PR #81.

1. clang: error: unknown argument: '-fvar-tracking-assignments-toggle'
clang does not support the -fvar-tracking-assignments-toggle flag;
This commit checks to see if the gcc compiler is in use before adding
this flag. For convenience, a compiler name variable has been added to
the SCons construction environment.

2. Building DRootGeom.cc against ROOT6 was failing because the -std=c++11
was missing. sbms.AddDANA(env) was moved to come before toggling the
-fvar-tracking-assignments flag to fix this.