This should fix second half of bug that caused program to crash in HDEVIO destructor. Initializes buffer pointers to NULL so if the file isan't opened and the constructor returns early, unitialized values aren't left in the pointers that the destructor then tries to free.
The first half of this bug is the CheckOpenable getting called
in the first place. It seems to be getting called after the last
valid source is finished, but it's unclear what source it is
trying to open. This is likely a bug in JANA, but I'm unable to
see where it is at the moment.