Remove bogus checks for TF1 in gDirectory. They are actually stored in gROOT->GetListOfFunctions().
However, it's far easier and safer for each thread to have their own
copy, rather than have one instance shared amongst threads.

That is because a user could modify the (e.g.) dE/dx cut, but keep the
same name, and it could supercede another user's cuts.

Although this invalidates the content of this global TF1 list, the fact
that this global list even exists is an abomination and an affront to humankind,
and should never be used.