Change what hypos are saved to the tree.
Before, if (e.g.) a given track with a given PID had 2 different RF beam bunches picked for it (in different combos), then BOTH of these hypos were saved separately to the "measured" hypo array. Now, only one is saved, and the PID info is in the combo-track section instead.

Same for neutrals.