Massive reduction in memory usage for kinematic fits.
This is achieved by no longer retaining the results of the dVEta and dV
covariance matrices. These contain the post-fit full error matrices for
the known parameters and the known+unknown params, respectively.

These were VERY large error matrices, who not only took up space but
also helped fragment the memory.

Also, change how the duplicate-checking history is reset, so that
histogram actions can be put in the combo pre-selection.