Overhaul handling of error matrices post-reconstruction.
1) Error matrices are now stored as floats (TMatrixFSym) instead of doubles (DMatrixDSym). This is
because they are only saved as floats in REST anyway, so this reduces
memory by 1/2.

2) Many, many, MANY error matrices are needed to store the results of
the kinematic fits. This is especially true in the analysis launches,
when running ~100 channels with 50+ threads, each with many combos. So,
to balance the load between threads, they are now stored in a resource
pool in DApplication. This pool is shared amongst all threads, and is
accessable anywhere in sim-recon. The matrices used for each thread are
recycled to the pool in the first matrix-request of a new event for that thread.
Using a memory pool also reduces memory fragmentation from excess
new/delete calls.

3) To use the memory pool, DKinematicData has been overhauled. It is no
longer the owner of matrix memory. This alone is a huge savings in
memory, as each time setErrorMatrix() was called prior, an entirely new
copy of the matrix was created and saved. This resulted in many more
new/delete/copy calls than needed, fragmenting the memory. Also, this
significantly reduces the complexity of the class, and allows me to
delete the god-awful goto statement.

4) These changes are post-reconstruction.  Meaning that while the
DBCALShower, DFCALShower, and DKinematicData objects have TMatrixFSym
matrices, they were calculated as doubles before being reduced to floats
at object creation.

5) Extraneous, dead code that needed to be fixed to be compatible with
these changes has instead been flat-out deleted.

Note: No attempt has been made to compile this yet.