Replace ugly map<step, step*> in blueprint factory with set<step*, comparator>.
The comparator uses the == operator of DParticleComboBlueprintStep, rather than the value of the pointers themselves.

This set makes it so that a new step can be quickly compared to previously-created ones to see whether or not it is unique.
If it is not unique, the memory for the new step is recycled, and the old step is used instead.

The problem with the previous implementation is that by using map<step, step*> it used a ton of memory.
Whereas using set<step*, comparator> uses significantly less memory.