Re-organize track/TOF/BCAL/FCAL/ST matching routines.
This consolidates code, removing inconsistencies.
Also, fixes to how matching to the ST geometry is done.
And, when swimming tracks, increase radius for aborting (used to be hard
cutoff at edge of TOF, but now is wider to take into account tracking
uncertainties).

Also, reorganize trackeff_missing plugin to make it easier to add
channels, and to be more consistent between them. And, place a cut
requiring that the missing track match to a hit in the ST, + either the
BCAL or BOTH the FCAL & TOF.