Do not extend binning of histograms.
1) This appears to be causing a crash sometimes when gROOT is trying to clean things up.
I don't know why this is happenning, but apparently it is.

2) This shouldn't be done anyway.  If you run this plugin over different files, you will
get different binning.  Then, when you try to merge the histograms, it has errors and
something unknown happens.  This feature is intended for drawing TTree contents in the TBrowser,
not for something that you know what it is ahead of time.  Thus, it has been deleted.