Multithreaded crashing fixed!
Problem was due to creating a resource on one thread, and recycling it
on another thread: race condition in the resource pool LOCAL pool.

How does this happen?  You create a charged-hypo on one thread, which
creates a DKinematicInfo object on that thread. You then recycle the
hypo, and another thread picks it up.  The other thread calls
Hypo::Reset(), which clears the DKinematicInfo.  Since it's stored in a
shared_ptr, it goes back to the thread it was created.

Solution: disable intra-thread sharing of objects that CONTAIN
shared_ptrs: The hypos & beam photons.

Yikes!