Fix issue with a line in this macro confusing the ROOT interpreter such that it renders other macros unable to draw to the RootSpy canvas. This bug seems to have appeared with the upgrade of ROOT to v6.10.00.
Remarkably, the fix is to put a space between the "p[2]"
and "exp" tokens. Specifically, change this:

gs=p[0]/sqrt(2.*3.1416)/p[2]*exp(-pow((x-p[1])/p[2],2)/2.);

to this:

gs=p[0]/sqrt(2.0*3.1416)/p[2] * exp(-pow((x-p[1])/p[2],2)/2.0);

the space after the "*" is not actually needed (I just put
it in for symmetry.) Interesting to note: this line causes
the problem when it is being read as part of a method
definition and not actually executed by the interpreter.
Go figure.