omega_skim plugin was flat-out broken prior to this.
It assumed that it was the FIRST analysis channel that was run, and
checking that channel to make the skim.  So if it was p2pi_hists
instead, it instead saved events that passed the RHO cuts. This is fixed
now.

Also, it crashed on events that weren't a physics trigger. This is fixed
now.